13-model_state_delete_a.py

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -13,6 +13,7 @@ if __name__ == "__main__":
         argv[1], argv[2], argv[3]), pool_pre_ping=True)
     Session = sessionmaker(bind=engine)
     session = Session()
-    stateUpdated = session.query(State).\
-        filter(State.name.contains('a')).delete()
-    session.commit()
+    if State.name.contains('a'):
+        stateUpdated = session.query(State).\
+            filter(State.name.contains('a')).delete()
+        session.commit()
